# Optimize sha256 for aarch64