Hi,

The sha256 is top hotspot for some applications, e.g. `Phase1` in `filecoin/lotus`, this patch optimizes it for aarch64.
It mainly remove the dependency between `ldr` and `add` in each four rounds by preloading `K` constants into NEON registers.
We also take advantages of `load/store pair` and `ld1` instructions to reduce the instructions executed.

Verified with repository `sha2` [1], test commands: `cargo +nightly test --features "asm"`, `cargo +nightly test --release --features "asm"`.

On our aarch64 server (core `tsv110`), we witnessed `11.90%` improvements with benches under `asm-hashes/sha2`:
```
# before
test bench_compress256 ... bench:          47 ns/iter (+/- 0) = 1361 MB/s
# after
test bench_compress256 ... bench:          42 ns/iter (+/- 0) = 1523 MB/s
```

[1] https://github.com/dgbo/hashes/tree/master/sha2