@tarcieri 2021-01-19
I can give this a try on an M1