@dgbo 2021-01-20
Thank you for watching this.

BTW, we run `lotus-bench` of `filecoin/lotus`, this PR reduce `10.5%` execution time of `Phase1` in this application.

Regards.