@newpavlov 2021-01-25
@tarcieri 
Can I leave this PR to you? Unfortunately I don't know ARM assembly well enough and do not have appropriate hardware to test the code.