@tarcieri 2021-01-25
Yep, been meaning to validate it on an M1. I'll see if I can take a look today.