@tarcieri 2021-01-25
Hrmm, interesting. The `sha2-asm` crate does not build on the M1 at all, either before or with this change:

```
   Compiling sha2-asm v0.5.4 (/Users/tony/asm-hashes/sha2)
The following warnings were emitted during compilation:

warning: src/sha256_aarch64.S:64:2: error: ADR/ADRP relocations must be GOT relative
warning:  adrp x2, .K
warning:  ^
warning: src/sha256_aarch64.S:64:2: error: unknown AArch64 fixup kind!
warning:  adrp x2, .K
warning:  ^
warning: src/sha256_aarch64.S:65:2: error: unknown AArch64 fixup kind!
warning:  add x2, x2, :lo12:.K
warning:  ^

error: failed to run custom build command for `sha2-asm v0.5.4 (/Users/tony/asm-hashes/sha2)`
```

I'd say this PR looks good to merge then, but separately we should track an M1 build failure.